requirements.txt
# pygame>=2.4.0,<2.9
pygame-ce>=2.3.1,<2.9
